python_service/requirements.txt
flask
numpy
scipy
librosa
sounddevice
soundfile
tensorflow
tensorflow_hub
requests
pyserial
